# Add support for typed internal pprof transfers