--- a/api/openapiv2/gen/phlare.swagger.json
+++ b/api/openapiv2/gen/phlare.swagger.json
@@ -1662,6 +1662,14 @@
         }
       }
     },
+    "v1PprofOutputMode": {
+      "type": "string",
+      "enum": [
+        "PPROF_BYTES",
+        "PPROF_TYPED"
+      ],
+      "default": "PPROF_BYTES"
+    },
     "v1PprofQuery": {
       "type": "object",
       "properties": {
@@ -1670,7 +1678,10 @@
           "format": "int64"
         },
         "stackTraceSelector": {
-          "$ref": "#/definitions/v1StackTraceSelector",
+          "$ref": "#/definitions/v1StackTraceSelector"
+        },
+        "outputMode": {
+          "$ref": "#/definitions/v1PprofOutputMode",
           "description": "TODO(kolesnikovae): Go PGO options."
         }
       }
@@ -1684,6 +1695,9 @@
         "pprof": {
           "type": "string",
           "format": "byte"
+        },
+        "typedPprof": {
+          "$ref": "#/definitions/googlev1Profile"
         }
       }
     },