--- a/cmd/pyroscope/help-all.txt.tmpl
+++ b/cmd/pyroscope/help-all.txt.tmpl
@@ -573,6 +573,8 @@ Usage of ./pyroscope:
     	Timeout for ingester client healthcheck RPCs. (default 5s)
   -querier.id string
     	Querier ID, sent to the query-frontend to identify requests from the same querier. Defaults to hostname.
+  -querier.internal-pprof-output-mode string
+    	The output mode for pprof queries, for transfers between internal components. Valid values are 'bytes' and 'typed'. (default "bytes")
   -querier.max-concurrent int
     	The maximum number of concurrent queries allowed. (default 4)
   -querier.max-flamegraph-nodes-default int