--- a/cmd/pyroscope/help.txt.tmpl
+++ b/cmd/pyroscope/help.txt.tmpl
@@ -157,6 +157,8 @@ Usage of ./pyroscope:
     	Run a health check on each ingester client during periodic cleanup. (default true)
   -querier.health-check-timeout duration
     	Timeout for ingester client healthcheck RPCs. (default 5s)
+  -querier.internal-pprof-output-mode string
+    	The output mode for pprof queries, for transfers between internal components. Valid values are 'bytes' and 'typed'. (default "bytes")
   -querier.max-flamegraph-nodes-default int
     	Maximum number of flame graph nodes by default. 0 to disable. (default 8192)
   -querier.max-flamegraph-nodes-max int