--- a/pkg/frontend/frontend.go
+++ b/pkg/frontend/frontend.go
@@ -114,6 +114,7 @@ type Limits interface {
 	QueryAnalysisEnabled(string) bool
 	SymbolizerEnabled(string) bool
 	QuerySanitizeOnMerge(string) bool
+	QueryInternalPprofOutputMode(string) string
 	validation.FlameGraphLimits
 }