--- a/pkg/test/integration/cluster/cluster_v2.go
+++ b/pkg/test/integration/cluster/cluster_v2.go
@@ -137,6 +137,9 @@ func (c *Cluster) v2Prepare(_ context.Context, memberlistJoin []string) error {
 					fmt.Sprintf("-query-backend.address=%s:%d", listenAddr, c.Components[compidx].grpcPort),
 				)
 			}
+			comp.flags = append(comp.flags,
+				"-querier.internal-pprof-output-mode=typed",
+			)
 		}
 
 		// handle metastore folders and ports