--- a/pkg/test/mocks/mockfrontend/mock_limits.go
+++ b/pkg/test/mocks/mockfrontend/mock_limits.go
@@ -378,6 +378,23 @@ func (_m *MockLimits) QuerySanitizeOnMerge(_a0 string) bool {
 	return r0
 }
 
+func (_m *MockLimits) QueryInternalPprofOutputMode(_a0 string) string {
+	ret := _m.Called(_a0)
+
+	if len(ret) == 0 {
+		panic("no return value specified for QueryInternalPprofOutputMode")
+	}
+
+	var r0 string
+	if rf, ok := ret.Get(0).(func(string) string); ok {
+		r0 = rf(_a0)
+	} else {
+		r0 = ret.Get(0).(string)
+	}
+
+	return r0
+}
+
 // MockLimits_SymbolizerEnabled_Call is a *mock.Call that shadows Run/Return methods with type explicit version for method 'SymbolizerEnabled'
 type MockLimits_SymbolizerEnabled_Call struct {
 	*mock.Call