--- a/pkg/validation/testutil.go
+++ b/pkg/validation/testutil.go
@@ -5,16 +5,17 @@ import (
 )
 
 type MockLimits struct {
-	QuerySplitDurationValue         time.Duration
-	MaxQueryParallelismValue        int
-	MaxQueryLengthValue             time.Duration
-	MaxQueryLookbackValue           time.Duration
-	QueryAnalysisEnabledValue       bool
-	QuerySanitizeOnMergeValue       bool
-	QueryAnalysisSeriesEnabledValue bool
-	MaxLabelNameLengthValue         int
-	MaxLabelValueLengthValue        int
-	MaxLabelNamesPerSeriesValue     int
+	QuerySplitDurationValue           time.Duration
+	MaxQueryParallelismValue          int
+	MaxQueryLengthValue               time.Duration
+	MaxQueryLookbackValue             time.Duration
+	QueryAnalysisEnabledValue         bool
+	QuerySanitizeOnMergeValue         bool
+	QueryInternalPprofOutputModeValue string
+	QueryAnalysisSeriesEnabledValue   bool
+	MaxLabelNameLengthValue           int
+	MaxLabelValueLengthValue          int
+	MaxLabelNamesPerSeriesValue       int
 
 	MaxFlameGraphNodesDefaultValue int
 	MaxFlameGraphNodesMaxValue     int
@@ -48,6 +49,9 @@ func (m MockLimits) QueryAnalysisSeriesEnabled(tenantID string) bool {
 }
 func (m MockLimits) QuerySanitizeOnMerge(tenantID string) bool {
 	return m.QuerySanitizeOnMergeValue
+}
+func (m MockLimits) QueryInternalPprofOutputMode(tenantID string) string {
+	return m.QueryInternalPprofOutputModeValue
 }
 func (m MockLimits) MaxFlameGraphNodesDefault(string) int { return m.MaxFlameGraphNodesDefaultValue }
 func (m MockLimits) MaxFlameGraphNodesMax(string) int     { return m.MaxFlameGraphNodesMaxValue }