… api/gen/proto/go/query/v1/query.pb.go